+++ packages/transformers-onnxruntime/src/tensor-ops.ts
