+++ packages/transformers/package.json
@@ -57,7 +57,7 @@
   "dependencies": {
     "@huggingface/jinja": "^0.5.6",
     "@huggingface/tokenizers": "^0.1.3",
-    "@huggingface/transformers-onnx": "workspace:^",
+    "@huggingface/transformers-onnxruntime": "workspace:^",
     "sharp": "^0.34.5"
   },
   "devDependencies": {
